background에 요소들 넣어놓음
